MULE-9072 Cleanup MuleSession. Move session handlers to compatibility module.
