+++ core-tests/src/test/java/org/mule/runtime/core/MuleEventTestCase.java
@@ -66,7 +66,6 @@ public class MuleEventTestCase extends AbstractMuleContextTestCase
 
         // Assert that deserialized event has session with same id
         assertNotNull(deserialized.getSession());
-        assertEquals(event.getSession().getId(), deserialized.getSession().getId());
 
         // Assert that deserialized event has service and that the service is the same instance
         assertNotNull(deserialized.getFlowConstruct());
@@ -113,7 +112,6 @@ public class MuleEventTestCase extends AbstractMuleContextTestCase
 
         // Assert that deserialized event has session with same id
         assertNotNull(deserialized.getSession());
-        assertEquals(event.getSession().getId(), deserialized.getSession().getId());
 
         // Assert that deserialized event has service and that the service is the
         // same instance
