+++ core/src/main/java/org/mule/runtime/core/VoidMuleEvent.java
@@ -18,6 +18,7 @@ import org.mule.runtime.core.api.context.notification.FlowCallStack;
 import org.mule.runtime.core.api.context.notification.ProcessorsTrace;
 import org.mule.runtime.core.api.processor.MessageProcessor;
 import org.mule.runtime.core.api.security.Credentials;
+import org.mule.runtime.core.api.security.SecurityContext;
 import org.mule.runtime.core.api.transformer.TransformerException;
 import org.mule.runtime.core.management.stats.ProcessingTime;
 
@@ -272,5 +273,17 @@ public class VoidMuleEvent implements MuleEvent
     public ProcessorsTrace getProcessorsTrace()
     {
         throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public SecurityContext getSecurityContext()
+    {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public void setSecurityContext(SecurityContext context)
+    {
+
     }
 }
