+++ transports/core-transports-tests/src/test/java/org/mule/compatibility/core/session/SessionPropertiesTestCase.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.session;
+package org.mule.compatibility.core.session;
 
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.sameInstance;
