+++ transports/core/src/main/java/org/mule/compatibility/core/session/NullSessionHandler.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.session;
+package org.mule.compatibility.core.session;
 
 import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.MuleException;
