+++ transports/core/src/main/java/org/mule/compatibility/core/session/SerializeAndEncodeSessionHandler.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.session;
+package org.mule.compatibility.core.session;
 
 import static org.mule.runtime.core.api.config.MuleProperties.MULE_SESSION_PROPERTY;
