+++ transports/core/src/main/java/org/mule/compatibility/core/transport/service/DefaultTransportServiceDescriptor.java
@@ -31,7 +31,7 @@ import org.mule.runtime.core.api.transformer.MessageTransformer;
 import org.mule.runtime.core.api.transformer.Transformer;
 import org.mule.runtime.core.config.i18n.CoreMessages;
 import org.mule.runtime.core.message.SessionHandler;
-import org.mule.runtime.core.session.SerializeAndEncodeSessionHandler;
+import org.mule.compatibility.core.session.SerializeAndEncodeSessionHandler;
 import org.mule.runtime.core.transaction.XaTransactionFactory;
 import org.mule.runtime.core.util.ClassUtils;
 import org.mule.runtime.core.util.CollectionUtils;
