+++ transports/core/src/main/resources/META-INF/mule-module.properties
@@ -29,6 +29,7 @@ artifact.export.classPackages=org.mule.compatibility.core,\
                               org.mule.compatibility.core.routing,\
                               org.mule.compatibility.core.routing.outbound,\
                               org.mule.compatibility.core.routing.requestreply,\
+                              org.mule.compatibility.core.session,\
                               org.mule.compatibility.core.transformer,\
                               org.mule.compatibility.core.transformer.simple,\
                               org.mule.compatibility.core.transport,\
