+++ transports/http/src/main/resources/META-INF/services/org/mule/runtime/transport/http.properties
@@ -5,7 +5,7 @@ message.receiver=org.mule.compatibility.transport.http.HttpMessageReceiver
 message.factory=org.mule.compatibility.transport.http.HttpMultipartMuleMessageFactory
 outbound.transformer=org.mule.compatibility.transport.http.transformers.ObjectToHttpClientMethodRequest
 response.transformer=org.mule.compatibility.transport.http.transformers.MuleMessageToHttpResponse
-session.handler=org.mule.runtime.core.session.SerializeAndEncodeSessionHandler
+session.handler=org.mule.compatibility.core.session.SerializeAndEncodeSessionHandler
 endpoint.builder=org.mule.compatibility.core.endpoint.UrlEndpointURIBuilder
 inbound.exchange.patterns=one-way, request-response
 outbound.exchange.patterns=one-way, request-response
